smp: Relax irqs_disable() warning in smp_call_function_single() The change that long-ago added the interrupts-disabled warning in smp_call_function_single() cited the following deadlock as the reason: 	CPU A				CPU B 	Disable interrupts 					smp_call_function() 					Take call_lock 					Send IPIs 					Wait for all cpus to acknowledge IPI 					CPU A has not responded, spin waiting 					for cpu A to respond, holding call_lock 	smp_call_function() 	Spin waiting for call_lock 	Deadlock			Deadlock
Such a deadlock is not possible however, in the event that
smp_call_function_single() is called for the currently-executing CPU.
While not a common usecase, relaxing the warning may simplify some
driver implementations.

The 'acpuclock-krait' driver in the MSM ARM sub-architecture provides
one example, where smp_call_function_single() is sometimes called from
a cpuidle path with interrupts disabled, but only ever from the target
CPU. Other callers of that same function may cross-call, but only do so
when interrupts are enabled.

Change-Id: I647b3b46fe6aa4dfb06f7750396986b80b92d700
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>